requirements.txt
opencv-python~=4.11.0.86
opencv-python-headless
numpy~=2.1.1
ultralytics~=8.3.96
matplotlib~=3.10.1

torch~=2.6.0
face-recognition~=1.3.0
ccxt~=4.4.77
pandas~=2.2.3
requests~=2.32.3
scikit-learn~=1.6.1
keras-nightly~=3.9.0.dev2025042303
torchvision~=0.21.0
pillow~=11.1.0
diffusers~=0.33.1
dlib~=19.24.6
yfinance~=0.2.55
keras~=3.9.2
ta~=0.11.0
python-telegram-bot~=22.0
tensorflow~=2.19.0